video-tools/requirements.txt
requests
openai
azure-cognitiveservices-speech
moviepy
audiocraft
